requirements.txt
Pillow==9.5.0
qrcode==7.4.2
streamlit==1.25.0
streamlit-option-menu==0.3.6
validators==0.20.0
